+++ backends/avx/ceed-avx-serial.c
@@ -43,8 +43,7 @@ static int CeedInit_Avx(const char *resource, Ceed ceed) {
 //------------------------------------------------------------------------------
 // Backend Register
 //------------------------------------------------------------------------------
-__attribute__((constructor))
-static void Register(void) {
-  CeedRegister("/cpu/self/avx/serial", CeedInit_Avx, 35);
+CEED_INTERN int CeedRegister_Avx_Serial(void) {
+  return CeedRegister("/cpu/self/avx/serial", CeedInit_Avx, 35);
 }
 //------------------------------------------------------------------------------