+++ backends/cuda-shared/ceed-cuda-shared.c
@@ -51,8 +51,7 @@ static int CeedInit_Cuda_shared(const char *resource, Ceed ceed) {
 //------------------------------------------------------------------------------
 // Register backend
 //------------------------------------------------------------------------------
-__attribute__((constructor))
-static void Register(void) {
-  CeedRegister("/gpu/cuda/shared", CeedInit_Cuda_shared, 25);
+CEED_INTERN int CeedRegister_Cuda_Shared(void) {
+  return CeedRegister("/gpu/cuda/shared", CeedInit_Cuda_shared, 25);
 }
 //------------------------------------------------------------------------------