+++ backends/opt/ceed-opt-blocked.c
@@ -64,8 +64,7 @@ static int CeedInit_Opt_Blocked(const char *resource, Ceed ceed) {
 //------------------------------------------------------------------------------
 // Backend Register
 //------------------------------------------------------------------------------
-__attribute__((constructor))
-static void Register(void) {
-  CeedRegister("/cpu/self/opt/blocked", CeedInit_Opt_Blocked, 40);
+CEED_INTERN int CeedRegister_Opt_Blocked(void) {
+  return CeedRegister("/cpu/self/opt/blocked", CeedInit_Opt_Blocked, 40);
 }
 //------------------------------------------------------------------------------