+++ backends/xsmm/ceed-xsmm-blocked.c
@@ -44,8 +44,7 @@ static int CeedInit_Xsmm_Blocked(const char *resource, Ceed ceed) {
 //------------------------------------------------------------------------------
 // Backend Register
 //------------------------------------------------------------------------------
-__attribute__((constructor))
-static void Register(void) {
-  CeedRegister("/cpu/self/xsmm/blocked", CeedInit_Xsmm_Blocked, 20);
+CEED_INTERN int CeedRegister_Xsmm_Blocked(void) {
+  return CeedRegister("/cpu/self/xsmm/blocked", CeedInit_Xsmm_Blocked, 20);
 }
 //------------------------------------------------------------------------------