+++ doc/sphinx/source/releasenotes.rst
@@ -19,6 +19,7 @@ New features
 * New HIP MAGMA backends for hipMAGMA library users: ``/gpu/hip/magma`` and ``/gpu/hip/magma/det``.
 * Julia and Rust interfaces added, providing a nearly 1-1 correspondence with the C interface, plus some convenience features.
 * New HIP backends for improved tensor basis performance: ``/gpu/hip/shared`` and ``/gpu/hip/gen``.
+* Static libraries can be built with ``make STATIC=1`` and the pkg-config file is installed accordingly.
 
 Performance improvements
 ^^^^^^^^^^^^^^^^^^^^^^^^